+++ src/textual/widgets/_list_view.py
@@ -143,14 +143,22 @@ class ListView(Vertical, can_focus=True, can_focus_children=False):
         return len(self.children)
 
     class Highlighted(Message, bubble=True):
-        """Emitted when the highlighted item changes. Highlighted item is controlled using up/down keys"""
+        """Emitted when the highlighted item changes. Highlighted item is controlled using up/down keys.
+
+        Attributes:
+            item (ListItem | None): The highlighted item, if there is one highlighted.
+        """
 
         def __init__(self, sender: ListView, item: ListItem | None) -> None:
             super().__init__(sender)
             self.item = item
 
     class Selected(Message, bubble=True):
-        """Emitted when a list item is selected, e.g. when you press the enter key on it"""
+        """Emitted when a list item is selected, e.g. when you press the enter key on it
+
+        Attributes:
+            item (ListItem): The selected item.
+        """
 
         def __init__(self, sender: ListView, item: ListItem) -> None:
             super().__init__(sender)