+++ tests/test_binding.py
@@ -6,25 +6,32 @@ from textual.binding import Bindings, Binding, BindingError, NoBinding
 
 BINDING1 = Binding("a,b", action="action1", description="description1")
 BINDING2 = Binding("c", action="action2", description="description2")
-BINDING3 = Binding("ctrl+d", action="action3", description="description3")
+BINDING3 = Binding(" d   , e ", action="action3", description="description3")
+BINDING4 = Binding("ctrl+d", action="action4", description="description4")
 
 
 @pytest.fixture
 def bindings():
-    yield Bindings([BINDING1, BINDING2, BINDING3])
+    yield Bindings([BINDING1, BINDING2, BINDING3, BINDING4])
 
+@pytest.fixture
+def more_bindings():
+    yield Bindings([BINDING1, BINDING2, BINDING3])
 
 def test_bindings_get_key(bindings):
     assert bindings.get_key("b") == Binding("b", action="action1", description="description1")
     assert bindings.get_key("c") == BINDING2
-    assert bindings.get_key("ctrl+d") == BINDING3
+    assert bindings.get_key("ctrl+d") == BINDING4
     with pytest.raises(NoBinding):
         bindings.get_key("control+meta+alt+shift+super+hyper+t")
+
+def test_bindings_get_key_spaced_list(more_bindings):
+    assert more_bindings.get_key("d").action == more_bindings.get_key("e").action
 
 def test_bindings_merge_simple(bindings):
     left = Bindings([BINDING1])
     right = Bindings([BINDING2])
-    eof = Bindings([BINDING3])
+    eof = Bindings([BINDING4])
     assert Bindings.merge([left, right, eof]).keys == bindings.keys
 
 def test_bindings_merge_overlap():