+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+### Added
+
+- Added `PRIORITY_BINDINGS` class variable, which can be used to control if a widget's bindings have priority by default. https://github.com/Textualize/textual/issues/1343
+
+### Changed
+
+- Renamed the `Binding` argument `universal` to `priority`. https://github.com/Textualize/textual/issues/1343
+- When looking for bindings that have priority, they are now looked from `App` downwards. https://github.com/Textualize/textual/issues/1343
+- `BINDINGS` on an `App`-derived class have priority by default. https://github.com/Textualize/textual/issues/1343
+- `BINDINGS` on a `Screen`-derived class have priority by default. https://github.com/Textualize/textual/issues/1343
+
+### Fixed
+
+- Fixed validator not running on first reactive set https://github.com/Textualize/textual/pull/1359
+- Ensure only printable characters are used as key_display https://github.com/Textualize/textual/pull/1361
 
 ## [0.6.0] - 2022-12-11
 
@@ -27,6 +44,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Fixed visibility not affecting children https://github.com/Textualize/textual/issues/1313
 - Fixed issue with auto width/height and relative children https://github.com/Textualize/textual/issues/1319
 - Fixed issue with offset applied to containers https://github.com/Textualize/textual/issues/1256
+- Fixed default CSS retrieval for widgets with no `DEFAULT_CSS` that inherited from widgets with `DEFAULT_CSS` https://github.com/Textualize/textual/issues/1335
+- Fixed merging of `BINDINGS` when binding inheritance is set to `None` https://github.com/Textualize/textual/issues/1351
 
 ## [0.5.0] - 2022-11-20